+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Add `gha-utils sponsor-label` command to label issues/PRs from GitHub sponsors.
 - Replace unmaintained `JasonEtco/is-sponsor-label-action` with `gha-utils sponsor-label`.
+- Change `gha-utils metadata` output path from positional argument to `-o/--output` option.
 - Fix `labels`, `workflows` and `bumpversion` commands fetching of `gha_utils/data/` content.
 - Fix race condition where `version-increments` job would skip due to missing tags.
 - Fallback to getting version from commits when tags aren't available yet.